move `impl pallet_staking::Config for Runtime` to separate file
